Added functionality to allow users to add tasks by pressing the Enter key.
An event listener was implemented on the text input field to detect when the Enter key is pressed. When the Enter key is pressed, the addTask() function is called, enabling the user to add a new task to the task list. This improvement provides a smoother and more intuitive user experience.